1/2 dino page + info widget
created half the dino info page, and a widget for info

+++ dino_app/lib/models/dinosaur.dart
@@ -1,9 +1,7 @@
-import 'dart:ffi';
-
 class Dinosaur{
 	String commonName, scientificName, era, 
     dietType, description, placeOfDiscovery, imageURL;
-  Float weight, height;
+  double weight, height;
 	
 	Dinosaur({
 		required this.commonName,
